Refactoring: Move tests into individual packages
